Move Atlas chat poll timeout to configuration file
The fewer behaviors that are hardcoded, the better. This commit moves the chat poll timeout to `config.atlas.poll` instead of leaving it hardcoded in the atlasChatStart command. The default value is 5s.

Also, reordered config properties to be alphabetical in Configuration.js.sample.

NOTE: This commit updates the expected Configuration.js properties and will require an update to that file in any existing installation.